data-engineering/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/data_engineering"]

[project]
name = "data-engineering"
version = "0.1.0"
description = "QuickPoll real-time analytics pipeline — PostgreSQL trigger functions, ETL, backfill"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3==1.34.0",
    "langchain>=1.0,<2.0",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.0,<2.0",
    "pandas==2.1.3",
    "pillow>=10.2.0",
    "psycopg2-binary==2.9.9",
    "pyconify>=0.1.6",
    "python-decouple==3.8",
    "python-pptx>=0.6.23",
    "rav>=0.4.3",
    "sqlalchemy==2.0.23",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.15.4",
]

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [".git", ".venv", "__pycache__", "build", "dist"]

[tool.ruff.lint]
select = ["E", "F", "B", "I", "UP", "SIM"]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
